[BK] Migrate batch 6 (custom nodejs build job)

+++ scripts/upload_nodejs_artifacts.sh
@@ -11,4 +11,9 @@ ARTIFACT_DIST_DIR="./workdir/dist"
 list_shasums_in_folder $ARTIFACT_DIST_DIR
 
 echo "--- Uploading build artifacts"
-gsutil cp -r "$ARTIFACT_DIST_DIR/*" gs://$BUCKET_NAME/$ARTIFACT_BASE_PATH
+if [[ "${DRY_RUN:-}" == "1" || "${DRY_RUN:-}" == "true" ]]; then
+  echo "Dry run enabled, skipping upload"
+  exit 0
+else
+  gsutil cp -r "$ARTIFACT_DIST_DIR/*" gs://$BUCKET_NAME/$ARTIFACT_BASE_PATH
+fi
